// packages/router-core/src/utils/pathToRegexp.d.ts
export interface MatchResult {
    path: string;
    index: number;
    params: Record<string, string>;
}
/**
 * A match is either `false` (no match) or a match result.
 */
export type Match = false | MatchResult;
/**
 * The match function takes a string and returns whether it matched the path.
 */
export type MatchFunction = (path: string) => Match;
// ...
export interface Key {
    name: string | number;
    prefix: string;
    suffix: string;
    pattern: string;
    modifier: string;
}